network_bot/requirements.txt
pep8==1.7.0
python-telegram-bot==5.3.0
SQLAlchemy==1.1.5
